pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "senior-mortgage-underwriting-system"
version = "1.0.0"
description = "Agentic AI Multi-Agent Mortgage Underwriting System"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0,<2.0.0",
    "langchain>=0.3.0,<2.0.0",
    "langchain-openai>=0.2.0,<2.0.0",
    "langchain-chroma>=0.1.2,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langchain-text-splitters>=0.3.0,<2.0.0",
    "chromadb>=0.5.0,<2.0.0",
    "pypdf>=4.0.0,<7.0.0",
    "streamlit>=1.38.0,<2.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
